03-cloud-functions-and-pub-sub-bui-jeremy/requirements.txt
Flask==2.3.3
google-cloud-storage==2.18.2
google-cloud-logging==3.1.2
google-cloud-pubsub==1.7.0
requests==2.25.1
functions-framework==3.0.0
six==1.16.0
